refactor: to reuse ManagementDetailInfoTable to where else (#8)
