+++ src/entities/home/setting/management/AdmissionManagementDetailEntities.ts
