+++ src/entities/home/setting/management/index.ts
@@ -1,3 +1,3 @@
 export * from "../../../../widget/ManagementDetailInfoTable";
-export { ManagementDetailButtons } from "./ManagementDetailButtons";
-export { uiEntities as managementDetailEntities } from "./managementDetailEntities";
+export { uiEntities as managementDetailEntities } from "./AdmissionManagementDetailEntities";
+export { AdmissionManagementDetailButtons as ManagementDetailButtons } from "./ManagementDetailButtons";
