+++ src/widget/ManagementDetailInfoTable.tsx
@@ -1,9 +1,4 @@
-import { Setting } from "@/shared/@types/setting";
 import Image from "next/image";
-import {
-  convertDataToTableEntity,
-  titleMapping,
-} from "../entities/home/setting/management/managementDetailEntities";
 
 const TableUnit = ({ title, data }: { title: string; data: string }) => {
   return (
@@ -19,19 +14,17 @@ const TableUnit = ({ title, data }: { title: string; data: string }) => {
 };
 
 export function ManagementDetailInfoTable({
-  admission,
+  data,
+  titleMapping,
 }: {
-  admission: Setting.GetAdmissionResponseDto;
+  data: { [key: string]: string };
+  titleMapping: { [key: string]: string };
 }) {
-  const data = convertDataToTableEntity(admission);
-
   return (
     <div className="grid h-full w-full grid-cols-2 justify-around gap-y-[27px]">
       {Object.keys(data).map((k) => {
         const key = k as keyof typeof data;
-        return (
-          <TableUnit key={key} title={titleMapping[key]} data={data[key]} />
-        );
+        return <TableUnit key={key} title={titleMapping[k]} data={data[key]} />;
       })}
     </div>
   );
