+++ src/widget/index.ts
@@ -1,3 +1,3 @@
+export { ManagementDetail } from "./AdmissionManagementDetail";
 export { Management, type ManagementState } from "./Management";
-export { ManagementDetail } from "./ManagementDetail";
 export { SideBar } from "./SideBar";
